Lab4/trip_planner/requirements.txt
streamlit==1.31.0
langchain==0.1.9
langchain-google-genai==0.0.9
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.1
